zensical:fix - new file permissions
`zensical new` will populate the target directory with files, by copying
them from the directory of the zensical python package.

When zensical is installed in as readonly, the permissions are copied
over, which means that the new files cannot be modified or deleted
(directories have the readonly permissions too).

This PR changes usage of:
- `shutil.copy` to `shutil.copyfile` and
- `shutil.copytree` to `shutil.copytree(copy_function=shutil.copyfile)`.

`shutil.copyfile` only copies the contents and produces files with 644.
`shutil.copytree` still produces readonly directories, so this PR also
includes a recursive walk of the new dir and changes directory
permissions to 755.

Signed-off-by: aljazerzen <aljaz@erzen.si>